another refactoring round

+++ mssql_to_redshift_data_transfer_tool/lib/aws.py
@@ -15,8 +15,8 @@ def _init_s3_client(aws_access_key_id, aws_secret_access_key):
     try:
         return boto3.client('s3', aws_access_key_id=aws_access_key_id,
                             aws_secret_access_key=aws_secret_access_key)
-    except ClientError:
-        raise MsSqlToRedshiftBaseException(ClientError)
+    except ClientError as cli_err:
+        raise MsSqlToRedshiftBaseException(cli_err)
 
 
 def _connect_to_redshift():
@@ -32,26 +32,36 @@ class Aws:
     def __init__(self):
         self.aws_access_key_id = getenv('aws_access_key_id')
         self.aws_secret_access_key = getenv('aws_secret_access_key')
-        self.s3_client = _init_s3_client(self.aws_secret_access_key, self.aws_secret_access_key)
+        self.s3_client = _init_s3_client(self.aws_access_key_id, self.aws_secret_access_key)
         self.redshift_conn = _connect_to_redshift()
 
-    def upload_to_s3(self, full_file_name, file_name):
+    def __repr__(self):
+        return self.redshift_conn
+
+    def upload_to_s3(self, dry_run, full_file_name, file_name):
+        if dry_run is True:
+            return
+
         try:
             transfer = S3Transfer(self.s3_client)
             transfer.upload_file(full_file_name, S3_BUCKET_NAME, S3_TARGET_DIR + "/" + file_name)
-        except S3UploadFailedError:
-            raise MsSqlToRedshiftBaseException(S3UploadFailedError)
+        except S3UploadFailedError as s3_upload_err:
+            raise MsSqlToRedshiftBaseException(s3_upload_err)
 
     def check_bucket(self):
         try:
             self.s3_client.head_bucket(Bucket=S3_BUCKET_NAME)
-        except ClientError:
-            raise MsSqlToRedshiftBaseException(ClientError)
+            return 0
+        except ClientError as cli_err:
+            raise MsSqlToRedshiftBaseException(cli_err)
 
     def disconnect_redshift(self):
-        self.redshift_conn.disconnect()
+        self.redshift_conn.close()
 
-    def copy_to_redshift(self, table_name, file_name):
+    def copy_to_redshift(self, dry_run, table_name, file_name):
+        if dry_run is True:
+            return
+
         try:
             cur = self.redshift_conn.cursor()
             cur.execute("begin;")
@@ -60,7 +70,8 @@ class Aws:
                                 + S3_TARGET_DIR + '/' \
                                 + file_name + "' credentials " \
                                 + "'aws_access_key_id=" + self.aws_access_key_id \
-                                + "; aws_secret_access_key=" + self.aws_secret_access_key + "' csv;"
+                                + ";aws_secret_access_key=" + self.aws_secret_access_key + "' csv;"
+
             cur.execute(copy_redshift_cmd)
             cur.execute("commit;")
         except psycopg2.Error as pse_err:
